# Conflicts:
#	database/pom.xml
#	service/src/main/java/cn/thinkfree/service/remote/CloudService.java
#	service/src/main/java/cn/thinkfree/service/remote/CloudServiceImpl.java

+++ service/src/main/java/cn/thinkfree/service/construction/vo/ConstructionOrderCommonVo.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class ConstructionOrderCommonVo {
 
     @ApiModelProperty("总页数")
-    private Integer countPageNum;
+    private long countPageNum;
 
     @ApiModelProperty("订单列表")
     private List<ConstructionOrderListVo> orderList;
